+++ src/ps/plone/mlstiles/support/cover/tests/test_listing_collection.py
@@ -3,14 +3,17 @@
 
 # python imports
 from mock import Mock
-try:
-    import unittest2 as unittest
-except ImportError:
-    import unittest
+import unittest
 
 # zope imports
-from collective.cover.testing import ALL_CONTENT_TYPES
-from collective.cover.tests.base import TestTileMixin
+try:
+    from collective.cover.testing import ALL_CONTENT_TYPES
+    from collective.cover.tests.base import TestTileMixin
+except ImportError:
+    ALL_CONTENT_TYPES = []
+
+    class TestTileMixin(object):
+        """Dummy class"""
 from plone import api
 from plone.app.testing import (
     TEST_USER_ID,
@@ -20,8 +23,10 @@ from plone.app.testing import (
 )
 
 # local imports
-from ps.plone.mlstiles.testing import PS_PLONE_MLSTILES_INTEGRATION_TESTING
-from ps.plone.mlstiles.tiles.listings import collection
+from ps.plone.mlstiles.testing import (
+    PS_PLONE_MLSTILES_INTEGRATION_TESTING,
+    skip_if_no_cover,
+)
 
 
 class ListingCollectionTileTestCase(TestTileMixin, unittest.TestCase):
@@ -29,32 +34,43 @@ class ListingCollectionTileTestCase(TestTileMixin, unittest.TestCase):
 
     layer = PS_PLONE_MLSTILES_INTEGRATION_TESTING
 
+    @skip_if_no_cover
     def setUp(self):
+        from ps.plone.mlstiles.support.cover import listing_collection
         super(ListingCollectionTileTestCase, self).setUp()
-        self.tile = collection.ListingCollectionTile(self.cover, self.request)
+        self.tile = listing_collection.ListingCollectionTile(
+            self.cover,
+            self.request,
+        )
         self.tile.__name__ = u'ps.plone.mlstiles.listings.collection'
         self.tile.id = u'test'
 
+    @skip_if_no_cover
     def test_interface(self):
         """Validate the tile implementation."""
-        self.interface = collection.IListingCollectionTile
-        self.klass = collection.ListingCollectionTile
+        from ps.plone.mlstiles.support.cover import listing_collection
+        self.interface = listing_collection.IListingCollectionTile
+        self.klass = listing_collection.ListingCollectionTile
         super(ListingCollectionTileTestCase, self).test_interface()
 
+    @skip_if_no_cover
     def test_default_configuration(self):
         """Validate the default tile configuration."""
         self.assertTrue(self.tile.is_configurable)
         self.assertTrue(self.tile.is_editable)
         self.assertTrue(self.tile.is_droppable)
 
+    @skip_if_no_cover
     def test_accepted_content_types(self):
         """Validate the accepted content types for the tile."""
         self.assertEqual(self.tile.accepted_ct(), ALL_CONTENT_TYPES)
 
+    @skip_if_no_cover
     def test_tile_is_empty(self):
         """Validate that the tile is empty by default."""
         self.assertTrue(self.tile.is_empty())
 
+    @skip_if_no_cover
     def test_delete_collection(self):
         """Validate behavior when the collection is removed."""
         obj = self.portal['mandelbrot-set']
@@ -84,32 +100,43 @@ class RecentListingsTileTestCase(TestTileMixin, unittest.TestCase):
 
     layer = PS_PLONE_MLSTILES_INTEGRATION_TESTING
 
+    @skip_if_no_cover
     def setUp(self):
+        from ps.plone.mlstiles.support.cover import listing_collection
         super(RecentListingsTileTestCase, self).setUp()
-        self.tile = collection.RecentListingsTile(self.cover, self.request)
+        self.tile = listing_collection.RecentListingsTile(
+            self.cover,
+            self.request,
+        )
         self.tile.__name__ = u'ps.plone.mlstiles.listings.recent'
         self.tile.id = u'test'
 
+    @skip_if_no_cover
     def test_interface(self):
         """Validate the tile implementation."""
-        self.interface = collection.IListingCollectionTile
-        self.klass = collection.RecentListingsTile
+        from ps.plone.mlstiles.support.cover import listing_collection
+        self.interface = listing_collection.IListingCollectionTile
+        self.klass = listing_collection.RecentListingsTile
         super(RecentListingsTileTestCase, self).test_interface()
 
+    @skip_if_no_cover
     def test_default_configuration(self):
         """Validate the default tile configuration."""
         self.assertTrue(self.tile.is_configurable)
         self.assertTrue(self.tile.is_editable)
         self.assertTrue(self.tile.is_droppable)
 
+    @skip_if_no_cover
     def test_accepted_content_types(self):
         """Validate the accepted content types for the tile."""
         self.assertEqual(self.tile.accepted_ct(), ALL_CONTENT_TYPES)
 
+    @skip_if_no_cover
     def test_tile_is_empty(self):
         """Validate that the tile is empty by default."""
         self.assertTrue(self.tile.is_empty())
 
+    @skip_if_no_cover
     def test_delete_collection(self):
         """Validate behavior when the collection is removed."""
         obj = self.portal['mandelbrot-set']
@@ -139,32 +166,43 @@ class FeaturedListingsTileTestCase(TestTileMixin, unittest.TestCase):
 
     layer = PS_PLONE_MLSTILES_INTEGRATION_TESTING
 
+    @skip_if_no_cover
     def setUp(self):
         super(FeaturedListingsTileTestCase, self).setUp()
-        self.tile = collection.FeaturedListingsTile(self.cover, self.request)
+        from ps.plone.mlstiles.support.cover import listing_collection
+        self.tile = listing_collection.FeaturedListingsTile(
+            self.cover,
+            self.request,
+        )
         self.tile.__name__ = u'ps.plone.mlstiles.listings.featured'
         self.tile.id = u'test'
 
+    @skip_if_no_cover
     def test_interface(self):
         """Validate the tile implementation."""
-        self.interface = collection.IListingCollectionTile
-        self.klass = collection.FeaturedListingsTile
+        from ps.plone.mlstiles.support.cover import listing_collection
+        self.interface = listing_collection.IListingCollectionTile
+        self.klass = listing_collection.FeaturedListingsTile
         super(FeaturedListingsTileTestCase, self).test_interface()
 
+    @skip_if_no_cover
     def test_default_configuration(self):
         """Validate the default tile configuration."""
         self.assertTrue(self.tile.is_configurable)
         self.assertTrue(self.tile.is_editable)
         self.assertTrue(self.tile.is_droppable)
 
+    @skip_if_no_cover
     def test_accepted_content_types(self):
         """Validate the accepted content types for the tile."""
         self.assertEqual(self.tile.accepted_ct(), ALL_CONTENT_TYPES)
 
+    @skip_if_no_cover
     def test_tile_is_empty(self):
         """Validate that the tile is empty by default."""
         self.assertTrue(self.tile.is_empty())
 
+    @skip_if_no_cover
     def test_delete_collection(self):
         """Validate behavior when the collection is removed."""
         obj = self.portal['mandelbrot-set']