+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add new base tiles which are independent of collective.cover.
+- Move collective.cover related tiles to new support package.
 
 
 0.3.1 (2017-08-24)